Moved button and modal for deletion to components/messaging directory
Relates #68
